pyproject.toml
[project]
name = "vspells-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Francesco Bertolaccini", email = "francesco.bertolaccini@trailofbits.com" }
]
requires-python = ">=3.12"
dependencies = [
    "logfire>=3.11.0",
    "mypy>=1.15.0",
    "pydantic-ai-slim[duckduckgo]>=0.0.45",
    "pydantic-ai[logfire]>=0.0.45",
    "pydantic-graph>=0.0.45",
    "ruff>=0.9.9",
]

[project.scripts]
vspells-ai-agent = "vspells_ai_agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
